docs: add some javadoc on E in hopes that it'll be visible in IDEs (#2714)
Closes: #2706

## Description

Add some documentation on *E*, in hopes that it will appear as hovertext
in IDEs. Also added an example and a link to the API Documentation.

### Security Considerations

No impact

### Scaling Considerations

No consequences.

### Documentation Considerations

No change to any interface. The goal is to add jsDoc to E.

### Testing Considerations

I don't know how to test this before it's deployed. In WebStorm, I was
able to see the comment when hovering on *E* in `no-shim.js`, but not
anywhere else.

### Compatibility Considerations

No effects

### Upgrade Considerations

None