fix(compartment-mapper): handle dynamically-required builtins properly; closes #2754
Dynamically-required builtins were going through the entire candidate-specifier-guess-'n'-check loop, which shouldn't happen.

Specifically, the call to a `MaybeReadNowFn` in this loop caused the original error as reported in #2754 because it uses a `FileUrlToPathFn`; the default implementation (Node.js' `url.fileURLToPath`) throws when provided `node:foo` _but does not throw_ when provided `foo`.

Also exports `WILDCARD_POLICY_VALUE` from `policy-format` so we can use it instead of magic strings all over the place.

Fixes #2754