chore: fix all usages of @module tags
Per [TypeDoc's documentation](https://typedoc.org/documents/Tags._module.html), a string following a `@module` tag _renames_ the module. JSDoc has this same behavior.

To that end, `@module`, if present, should _not_ be at the _top_ of a docstring (which is used at the description) unless we're in the business of renaming modules for documentation purposes (which we are welcome to do later, if we want).

* * *

There are many ways to do this, but I did it using [ripgrep](https://github.com/BurntSushi/ripgrep) and `sponge` from [moreutils](https://joeyh.name/code/moreutils/) (which—TIL—also contains a utility named `pee`):

```sh
for file in packages/*/src/**/*.{t,j}s; do
  rg -P '(?<=\/\*\*\n\s\*)\s@module(\s[\s\S]+?)(?=\n\s\*\/)' -U \
    --passthru -r '$1
*
* @module' $file | sponge $file
done
```

- `-P`: use PRCE2 (for the lookbehind)
- `-U`: enable multiline regex
- `--passthru`: dump the entire file
- `-r`: replace