fix(compartment-mapper): fix crawling (#2769)
The `node_modules`-crawler in `@endo/compartment-mapper` is currently unstable. When there are multiple paths to a module, the "shortest path" to that module remains the first path that was found.

The first path that is found is non-deterministic.  As a result, a `CompartmentDescriptor.path` value can differ from run to run.

This fixes the problem: when a package location in the graph is encountered for the _second_ time (and every time thereafter), we recursively (and its dependency tree, if that too needs updating) check if it still has the shortest path, and update paths as appropriate within the graph.

A fixture in `packages/compartment-mapper/test/fixtures-shortest-path` shows an example of the structure necessary to trigger the original problem.  Previous to the change, this test would typically fail within 20 iterations, thus exhibiting non-determinism.