feat(ses): ArrayBuffer.prototype.transferToImmutable (#2400)
Closes: #XXXX
Refs: #XXXX

## Description

Now that the Immutable ArrayBuffer proposal is at 2.7 and looks likely
to proceed to 3, and then on to 4 and the official language, we are now
allowing ourselves to build on the shim to emulate it already being in
the language. In this PR ses imports the shim (not just the ponyfill)
and then permits it's additions as harmless for ses users.

### Security Considerations

We carefully designed the proposal to maintain our high security
standards. Likewise, in our shim, an Immutable ArrayBuffer encapsulates
a normal mutable ArrayBuffer. But it is careful not to mutate it or to
let it escape. The shim is low fidelity in terms of what it must
practically omit --- TypedArrays and DavaView backed by Immutable
ArrayBuffers. But these omissions do not raise security hazards, only
compat hazards.

### Scaling Considerations

Because Immutable ArrayBuffers obviate the need for defensive copying
and support safe zero-copy sharing, it will help scalability.

- [ ] When on XS, we should somehow arrange to use its high-speed
high-fidelity highly tested Immutable ArrayBuffer implementation. We
should also figure out our migration strategy as other engines of
interest ship their own standards compliant implementations. All these
steps will help scaling immensely. But it is unclear how to handle the
transition.

### Documentation Considerations

Mostly, we can just point at the docs for Immutable ArrayBuffer itself,
currently in the proposal and soon in the JS standard.

### Testing Considerations

The XS native Immutable ArrayBuffer implementation is also highly tested
by the test262 tests in the PR they submitted, which will also test
other native implementations.

- [ ] Our shim is not highly tested. We should somehow test it against
the relevant subset of those submitted test262 tests.

### Compatibility Considerations

The shim is low fidelity in terms of what it omits from the standard.
The shim is also low fidelity in that it cannot hide an internal shim
implementation artifact, the hidden prototype that all and only
Immutable ArrayBuffers share. A correct implementation of the proposal
will have no such extra prototype object.

### Upgrade Considerations

- [x] Update `NEWS.md` for user-facing changes.

---------

Co-authored-by: LeoTM <1881059+leotm@users.noreply.github.com>