feat(pass-style,marshal): ByteArray, a new binary Passable type (#1538)
Closes: #1331 
Refs: https://github.com/ocapn/ocapn/issues/5#issuecomment-1492778252
#2414

## Description

Recognize Immutable ArrayBuffers, when also frozen, as the new
pass-style `byteArray`. Add a branch for each of the dispatches on
pass-style that must now understand byte-arrays as well. As of this PR,
many of those additional branches are stubbed out with "not yet
implemented" errors as placeholders.

### Security Considerations

The underlying shim implementation, if we're still using that, uses a
normal mutable ArrayBuffer internally, but safely encapsulates it so
nothing outside the shim can mutate it.

### Scaling Considerations

The underlying shim implementation helps us handle bulk binary data in a
largely no-copy or minimal-copy manner.

### Documentation Considerations

Will add a new pass-style, so everywhere we enumerate all the
pass-styles, we'll need to add byte-array.

### Testing Considerations

- [ ] TODO tests. For each stubbed out case, there instead should be a
`test.failing` test.

### Compatibility Considerations

Old code that dispatched only on the pass-styles it knew about will not
correctly handle passables that contain byte-arrays.

The shim cannot reasonably support TypedArrays of DataViews backed by
Immutable ArrayBuffers.

### Upgrade Considerations

- [ ] TODO BREAKING.md
- [ ] TODO NEWS.md