fix(pass-style): `isAtom` must accept non-passables (#2803)
Closes: #XXXX
Refs: #2791 

## Description

Before #2791 , the Passables that correspond with the OCapN "Atom" were
called "primitives", and were originally aligned with JS primitive
value. Then two exceptions were introduced: ByteArrays and Symbols are
both OCapN Atoms. But ByteArrays are not and cannot be JS primitive
values, and we expect to switch our JS implementation of Symbols so they
are represented in JS as objects rather than JS symbols.

#2791 changed our terminology from "Passable Primitives" to "Atom". It
also introduced `isAtom` and `assertAtom` for testing at this level of
abstraction. Old calls to `isObject` that were testing JS primitiveness
that were really about atomness were changed to use the new predicates.

But it was not yet the drop in replacement that I thought it would be.
`isObject`, operating at the JS level, always returned a boolean and
never threw. `isAtom` on a non-Passable threw. This broke some uses of
`deeplyFulfilled` where the top value happened to be a non-frozen
promise, which was therefore non-Passable. Because `deeplyFulfilled`
immediately unwrapped that top level promise to its settlement, it
remains ok for `deeplyFulfilled` to be tolerant of a top level
non-Passable promise.

This PR fixes `isAtom` so that it always returns a boolean. If given a
non-Passable, `isAtom` will now return `false`.

### Security Considerations

none
### Scaling Considerations

none
### Documentation Considerations

`deeplyFulfilled`'s tolerance of a top level non-Passable promise is
surprising and irregular. However, because it had been tolerant, code
accumulated that happened to make use of that tolerance, probably
without knowing that it was doing something irregular.

### Testing Considerations

- [ ] expand `atom.test.js`

### Compatibility Considerations

Perhaps a better plan would be to fix all callers of `deeplyFulfilled`.
But we're not doing that in this PR. If we ever do, we'll have to
wrestle with the compat problems then. This PR merely restores the
compat that #2791 unknowingly broke.

### Upgrade Considerations

#2791 has not been released yet. Assuming it only gets released together
with this PR correcting it, there should be no external compat or
upgrade considerations.