fix(compartment-mapper): handle invalid peer dependencies
This adds dependency names found in `peerDependenciesMeta` which are otherwise unspecified in other dependency fields to the list of all dependencies so that a `CompartmentDescriptor` will be created if the package exists.

We will prefer any spec already provided in a dependency field instead of the default (which is `*` for lack of a better value).