fix(compartment-mapper): handle invalid peer dependencies
This adds dependency names found in `peerDependenciesMeta` which are otherwise unspecified in other dependency fields to the list of all dependencies so that a `CompartmentDescriptor` will be created if the package exists.

While entries in `peerDependenciesMeta` are meaningless to `npm` without a corresponding entry in `peerDependencies`. The reasons for providing such an invalid `package.json` are historical; previous to `npm@7`, there was no way to declare an _optional_ peer dependency.

That said, the _intent_ of such a configuration is to declare a package as an _optional peer dependency_; thus, `@endo/compartment-mapper` will attempt to honor it.

This changes `graphPackage()` so that it no longer stores unused references to the _specification_ of any given dependency.