fix(ses): avoid depth-first loading sequence in loadNow to prevent getting stuck in cycles (#2804)
## Description

It took a while, but we've encountered case that would first go into a
dynamic require (therefore trigger compartment-mapper using importNow)
and proceed to import a codebase with a cycle in cjs modules.

The minimal implementation of enqueueJob that we used in importNow was
causing the synchronous case for `memoizedLoadWithErrorAnnotation` to
become depth-first and getting into cycles before anything got
successfully memoized.

The new `syncJobQueue` guarantees a breadth-first order. 

### Security Considerations

none

### Scaling Considerations

`current` and `next` in `syncJobQueue` get swapped and each layer of the
module dependency tree gets processed as a separate step, which should
be reasonably memory efficient. I could eliminate recursion from
`drainQueue` altogether, but that seems a premature optimization.

### Documentation Considerations

It's a fix

### Testing Considerations

The testcase is minimal and gets to stack limit fairly quickly when
failing.

### Compatibility Considerations

I noticed signatures of `import` and `importNow` differ. I don't recall
us making that decision, but `await import` gives us `{ namespace }`
while `importNow` gives us `namespace` directly. Is that some
future-proofing I don't remember?

### Upgrade Considerations

N/A