feat(panic): ponyfill `panic` from "Don't Remember Panicking" proposal (#2815)
Closes: #XXXX
Refs: https://github.com/Agoric/agoric-sdk/pull/11173
https://github.com/tc39/proposal-oom-fails-fast

## Description

Adds `@endo/panic` as an imperfect ponyfill of the `panic` function from
the [Don't Remember
Panicking](https://github.com/tc39/proposal-oom-fails-fast) tc39 draft
proposal. Please see the `README.md` in `@endo/panic` for example use
and details.

### Security Considerations

Felipe Natal reported a security weakness in Agoric's handling of
alleged all-or-nothing transactions, especially in the context of
transfer or erights. While we do not know of any practical way for an
outside party to trigger a throw during the execution of the "commit
phase" (see the README.md), if such an error could be induced in the
absence of a reliable `panic` operation, execution might proceed with
the transfer partly done, causing either the disappearance of erights or
their duplication, both of which would violate conservation.

This `panic` ponyfill is imperfect because it cannot reliably and
portably immediately exit on all platforms. However, it provides a hook
that `@agoric/swingset/liveslots` will use so that `panic` in a vat
running on SwingSet will reliably immediately exit, preserving
transactional integrity for that environment.

But in general, the hazard will still remain.

### Scaling Considerations

Under SwingSet, we'd like a vat `panic` to immediately terminate the vat
incarnation but only suspend the vat, so that the vat can be upgraded
from a state that still had transactional integrity to recover and
continue correct execution.

### Documentation Considerations

We should rely on the draft tc39 proposal to provide most of the needed
documentation for the "real" `panic`.

The README.md of `@endo/panic` does provide adequate documentation of
how the ponyfill differs from the "real" one.

### Testing Considerations

It's hard to write an ava test that tests immediate exit, since (in the
same agent) there is no longer an ava to observe that the `exit`
happened. Instead, this PR finds adequate indirect measures to test
whether this ponyfill does what it is supposed to do -- as described in
its README.

### Compatibility Considerations

Just introducing the ponyfill should have no compat considerations.
Later PRs that start using it, such as draft
https://github.com/Agoric/agoric-sdk/pull/11173, may well have their own
compat issues. Looking forward, as the tc39 proposal evolves, we'd like
to evolve the ponyfill to best match it. Once the proposal settles down,
we'd like to add a shim to the ponyfill to emulate the settled proposal.

### Upgrade Considerations

To support vat upgrade under liveslots, as noted, liveslots will need to
react to any fatal termination including `panic` by terminating only the
vat incarnation, suspending the vat, enabling the vat to upgrade and
continue from a consistent state in a new incarnation. Liveslots does
not yet have the support needed for such vat suspension.