refactor(pass-style,...): abstract symbol making to prepare for flip (#2793)
Closes: #XXXX
Refs: #2792 #2777

## Description

This PR is essentially a pure refactor to set things up for upcoming
substantive PRs like #2792 . The main one will change the passable
symbol representation, along the lines of #2777 but likely different.
This is for symbols used for purposes passing, or test counterexamples
to being passable. All other symbol uses should not change.

The change from this PR is
- change calls to `Symbol.for('foo')` for making a passable symbol, to
`passableSymbolForName('foo')`
- change calls to `Symbol('foo')` for making a non-passable symbol for
testing purposes, to `unpassableSymbolForName('foo')`

### Security Considerations

Assuming we changed each symbol-making callsite from the old to the
correct new one, and assuming that we did not alter the code that was
making symbols for other reasons, none.

### Scaling Considerations

Using registered symbols like `Symbol.for('foo')` is a semantically
perfect binding to the pass-style and ocapn category `'symbol'`. However
its scaling properties are disastrous. In some js systems like v8,
unused registered symbols are not garbage collected. This is why we're
going to switch passable symbol representation. However this PR does
none of that yet. It merely sets the stage. This PR should not cause any
change in scaling.

### Documentation Considerations

The need to pass symbols is rare and obscure, and most users will not
need to be aware of this. However, any reference documentation should
clearly explain when to use the new abstract symbol-creation functions,
and when not to.

### Testing Considerations

As a pure refactor, the main concern is that all the old tests still
pass. Any substantive new tests or test changes will be downstream, in
the PRs that this PR enables.

### Compatibility Considerations

Assuming we changed each symbol-making callsite from the old to the
correct new one, and assuming that we did not alter the code that was
making symbols for other reasons, none.

### Upgrade Considerations

As a pure refactor, none. Beyond that, when we change the JS
representation of a passable symbol, we will not change the wire
representation in any of our encodings. Thus, durable state saved using
the representation will, at that time, revive into the new
representation, so still none. Likewise with messages sent from code on
one side of the upgrade and received by code on the other side of the
upgrade.