fix(immutable-arraybuffer): capture structuredClone early (#2853)
Closes: #XXXX
Refs: https://github.com/endojs/endo/pull/2785

## Description

The Immutable ArrayBuffer shim, on non-Hermes, relies on the existence
of *either* a global `structuredClone` or `Array.prototype.transfer`.
However, I forgot to capture early the global `structuredClone` into a
module-local one, and to then use that repeatedly. MetaMask scuttling
deleted the global `structuredClone`, running headlong into my mistake
on platforms (Node 18 and Node 20) that do not also have
`Array.prototype.transfer`.

This PR fixes that mistake, capturing `structuredClone` early, and then
using the resulting module-local one.

### Security Considerations

The bug this PR fixes was a security hazard.

### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

In theory, we should add scuttling as an orthogonal dimension of our
platform testing matrix. However that is rather expensive for minor
benefit.

### Compatibility Considerations

The bug created compat problems that this PR fixes.
### Upgrade Considerations

none