feat(compartment-mapper): support mapping of undiscoverable packages
This adds the `additionalPackageLocations` and `additionalPackageDetails` options to `mapNodeModules`. The former is a mapping of module location to one or more additional module locations. The latter is a data structure `mapNodeModules` will populate, which can then be provided to `captureFromMap` (which is in turn provided to `makeImportHookMaker()`.

This is intended to be used when dynamic requires prevent `mapNodeModules` & `captureFromMap` from otherwise creating a complete `CompartmentMapDescriptor` due to its ignorance of what would be dynamically required.  Archival is not a use-case.

Also:

- Fixed signature of `chooseModuleDescriptor`
- Tweaked ESLint config to ignore my new fixture
- Add options type for `translateGraph`
- Fixed deprecation notice for `compartmentMapForNodeModules`; it was previously on the options object, but it should only be on the export.
- `graphPackages` now accepts an optional `Graph` object to be able to re-use the same graph between runs