fix(marshal): fix consequence of typo (#2841)
## Description

The typo was that in a NEWS.md file, I said "Nest release" instead of
"Next release". The consequence is that the release that happened since
then missed this one, so it is too late to fix it simply by revising to
"Next release". Instead, I manually revised it to what I imagine the
release process would have renamed it to.