fix(immutable-arraybuffer): unify shim to work on more platforms (#2855)
Closes: #XXXX
Refs: #2785 #2399

## Description

The Immutable ArrayBuffers shim can only emulate `transferToImmutable`
on platforms with either `structuredClone` or
`ArrayBuffer.prototype.transfer`. Otherwise, the shim would refuse to
initialize. Without Immutable ArrayBuffers, we cannot initialize `ses`,
and `pass-style` cannot implement the new `byteArray` Passable type,
which `marshal` and `patterns` now depends on.

Unfortunately, it turns out that some platforms we want to support do
not have `structuredClone` or `ArrayBuffer.prototype.transfer`. The
first we encountered was Hermes, that also does not have classes with
private fields. #2785 makes all these work on Hermes, by supporting
`sliceToImmutable` instead, and using WeakMaps rather than private
fields. On the theory that Hermes was a one-off, #2785 sniffed to see if
it was on Hermes, only only then used the alternate shim.

In turns out Hermes was not a one-off. Some version of JSC of concern
(TODO which ones?) also have neither `structuredClone` or
`ArrayBuffer.prototype.transfer`. So this PR unifies the two shim
strategies into one which uses the WeakMap technique, which works
everywhere, and emulating `sliceToImmutable`, which we can do
everywhere. It still sniffs for `structuredClone` or
`ArrayBuffer.prototype.transfer`, and only emulates
`transferToImmutable` if one is present. Thus, we sniff only for
features not platforms, and emulate `sliceToImmutable` on all platforms.

Other PRs (TODO: which ones?) already ensure that `pass-style`,
`marshal`, and `patterns` are fine with the absence of
`transferToImmutable`.

### Security Considerations

A singe piece of code with small feature detects is simpler, and thus
more secure, than maintaining two alternate pieces of code, each for a
different set of platforms.

### Scaling Considerations

On most platforms, a class with private fields is probably more
efficient than a WeakMap. This PR sacrifices that optimization in
exchange for unification.

### Documentation Considerations

Less to document.

### Testing Considerations

The "unified" tests are just the union of the old tests, and so have two
problems
- [ ] redundancy, which needs to be removed
- [ ] contains unconditional `transferToImmutable` tests that won't work
on some platforms. These specific tests need to be made conditional,
probably using the same `test.skip` trick some tests use for
`Array.prototype.transfer`.

### Compatibility Considerations

The goal is to make the shim compat with more platforms. On platforms
that worked before this PR, this PR should be a pure refactor, and so
without any compat hazards.

### Upgrade Considerations

- [x] Update `NEWS.md` for user-facing changes.

---------

Co-authored-by: Christopher Hiller <boneskull@boneskull.com>