refactor(pass-style): faster `isObject` (#2860)
Closes: #XXXX
Refs: https://github.com/Agoric/agoric-sdk/pull/11497 

## Description

Experiment with possible endo speedups. This was originally motivated to
try for fewer benchmark errors in
https://github.com/Agoric/agoric-sdk/pull/11497 vs
https://github.com/Agoric/agoric-sdk/pull/11459 . However, the relevant
bottleneck for those, while still not diagnosed, is not addressed or
affected by this PR.

Nevertheless, according to @gibson042 's comment at
https://github.com/endojs/endo/pull/2860#pullrequestreview-2951292178
below

> This change has minor impact in V8, but Object(value) must be very
slow in XS because the speedup there is significant.

Since we care most of all about performance in XS, this change is worth
it.


### Security Considerations

none
### Scaling Considerations

No change to complexity measure, so no change to "scaling" per se. But
more colloquially, we include in "scaling" a significant performance
improvement.

### Documentation Considerations

In light of @gibson042 's suggestion below, this PR will deprecate
`isObject`, export `isPrimitive` as the inverted replacement, because
the name is clearer, and will switch all current callers of `isObject`
to call `isPrimitive` instead. This should be reflected in the generated
typedoc commentary.

### Testing Considerations

None.

### Compatibility Considerations

Because we're only deprecating `isObject`, and because we make no change
to its observable behavior, none.

### Upgrade Considerations

- [ ] Update `NEWS.md` with `isPrimitive` and the deprecation of
`isObject`.