refactor(base64): minor improvements (#2870)
Closes: #XXXX
Refs: #2844 

## Description

https://github.com/endojs/endo/pull/2844 currently depends on
`@endo/base64` for base64 encoding and decoding. @kriskowal asked that
it be changed to use a hex encoding instead, for which changes to
`@endo/base64` would be irrelevant. So I'm capturing in this PR the
minor improvement I made to that package while working on #2844 .

### Security Considerations

exporting mutable data (`monodu64`) from a module is a security hazard.
It is not exported from the package, so this is only a hazard, not a
vulnerability. In this case there was zero reason not to freeze
`monodu64`, so this PR does, leaving us one less security hazard to
worry about.

### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

Aside from `monodu64`, the other changes to this PR is to expand the
test in ways I noticed while working on #2844 .

### Compatibility Considerations

none

### Upgrade Considerations

none

---------

Co-authored-by: Richard Gibson <richard.gibson@gmail.com>